Added to filter, map, zip testing for mutating the underlying iterables
